+++ main.jl
@@ -4,6 +4,7 @@ include("Parameters.jl")
 using .Parameters
 using Distributed
 addprocs(prod(nprocs_perdim),topology=:all_to_all,lazy=true)
+using DistributedArrays
 
 include("IC.jl")
 include("Energy.jl")
@@ -45,31 +46,51 @@ using PProf
     using .IndexMap
     using .Time_Evolution
     using .Constraints_Conserveds
+    using DistributedArrays
 end
 
 
 #--Initilize the Global Lattice Arrays 
     #Standards:  ϕ and ψ are in 2D lattice // Z can be on the flattened 1D N^2 lattice or native 2D lattice
-    const ϕ_gl = OffsetArray(im*zeros(Nx, Ny),lx:rx,ly:ry)
-    const ψ_gl = OffsetArray(zeros(Nx, Ny),lx:rx,ly:ry)
-    const ZED_gl = OffsetArray(zeros(Nx,Ny),lx:rx,ly:ry)
+    # const ϕ_gl = OffsetArray(im*zeros(Nx, Ny),lx:rx,ly:ry)
+    # const ψ_gl = OffsetArray(zeros(Nx, Ny),lx:rx,ly:ry)
+    # const ZED_gl = OffsetArray(zeros(Nx,Ny),lx:rx,ly:ry)
+
+    const ϕ_gl = im*zeros(Nx, Ny)
+    const ψ_gl =    zeros(Nx, Ny)
+
+    const ϕ = im.*dzeros(Nx_padd,Ny_padd,2)
+    const ψ =     dzeros(Nx_padd,Ny_padd,2)
+    const dϕdt = im.*dzeros(Nx,Ny,2)
+    const dψdt =     dzeros(Nx,Ny,2)
+    const Z = DArray((Nx_padd,Nx,Ny_padd,Ny,2),workers(),[nprocs_perdim[1],1,nprocs_perdim[2],1,1]) do I
+        im*zeros(length(I[1]),length(I[2]),length(I[3]),length(I[4]),length(I[5]))
+    end
+    const dZdt = DArray((Nx,Nx,Ny,Ny,2),workers(),[nprocs_perdim[1],1,nprocs_perdim[2],1,1]) do I
+        im*zeros(length(I[1]),length(I[2]),length(I[3]),length(I[4]),length(I[5]))
+    end
+
+
+    @sync @distributed for p in workers()
+        
+    end
 
 
 function run_ev()
 
 #--Initilize the Chunk Field Arrays
-    @everywhere workers() begin 
-        ϕ =    im*zeros(Nx_loc+padding_size, Ny_loc+padding_size, 2)
-        ψ =       zeros(Nx_loc+padding_size, Ny_loc+padding_size, 2)
-        dϕdt = im*zeros(Nx_loc, Ny_loc, 2) #!it looks like these don't need the padding.
-        dψdt =    zeros(Nx_loc, Ny_loc, 2) #!but easier for "for" loops when ϕ and dϕdt are in the same one (see dZdt)
-        #2-index Z
-        # Z =    Array{ComplexF64,3}(undef, Nx^2,Ny^2,2) #This way seems to be faster and memory friendely for very large complex arrays. 
-        # dZdt = Array{ComplexF64,3}(undef, Nx^2,Ny^2,2)
-        #4-index Z
-        Z =    Array{ComplexF64,5}(undef, Nx_loc+padding_size, Nx, Ny_loc+padding_size, Ny, 2)
-        dZdt = Array{ComplexF64,5}(undef, Nx_loc, Nx, Ny_loc, Ny, 2) #!BUT for this one it might be a huge overhead!(see above red)
-    end
+    # @everywhere workers() begin 
+    #     ϕ =    im*zeros(Nx_loc+padding_size, Ny_loc+padding_size, 2)
+    #     ψ =       zeros(Nx_loc+padding_size, Ny_loc+padding_size, 2)
+    #     dϕdt = im*zeros(Nx_loc, Ny_loc, 2) #!it looks like these don't need the padding.
+    #     dψdt =    zeros(Nx_loc, Ny_loc, 2) #!but easier for "for" loops when ϕ and dϕdt are in the same one (see dZdt)
+    #     #2-index Z
+    #     # Z =    Array{ComplexF64,3}(undef, Nx^2,Ny^2,2) #This way seems to be faster and memory friendely for very large complex arrays. 
+    #     # dZdt = Array{ComplexF64,3}(undef, Nx^2,Ny^2,2)
+    #     #4-index Z
+    #     Z =    Array{ComplexF64,5}(undef, Nx_loc+padding_size, Nx, Ny_loc+padding_size, Ny, 2)
+    #     dZdt = Array{ComplexF64,5}(undef, Nx_loc, Nx, Ny_loc, Ny, 2) #!BUT for this one it might be a huge overhead!(see above red)
+    # end
 
 
     #--Information about the run
@@ -93,14 +114,27 @@ function run_ev()
     ioϕ=open("data/initial_phi.dat","w")
     ioψ=open("data/initial_psi.dat","w")
 
+    
 
 
     #----Initial Conditions----#
-    @time initialConditions!(ϕ_gl,ψ_gl)
+    @time initialConditions!(ϕ,ψ,Z,dϕdt,dψdt,dZdt,ϕ_gl,ψ_gl)
 
     #Record initial conditions
+    open("data/initial_phi.dat","w") do ioϕ
+    open("data/initial_psi.dat","w") do ioψ
+    # for p in workers()
+    #     lx_p, rx_p, ly_p, ry_p = distChunker(p)
+    #     ψ_gl[lx_p:rx_p,ly_p:ry_p] .= @fetchfrom p localpart(ψ)[1+padd:Nx_loc+padd,1+padd:Ny_loc+padd,1]
+    #     ϕ_gl[lx_p:rx_p,ly_p:ry_p] .= @fetchfrom p localpart(ϕ)[1+padd:Ny_loc+padd,1+padd:Ny_loc+padd,1]
+    # end
     writedlm(ioϕ,ϕ_gl[:,:])
-    writedlm(ioψ,ψ_gl[:,:])
+    # writedlm(ioψ,ψ_gl[:,:])
+    writedlm(ioψ,ψ[:,:,1]) #! this doesn't work for me since there are paddings 
+                            #!either remove paddings on the plotting side or just use a separate "global" field array
+    end
+    end
+
 
 
     # #!Turn on if you wanna check constraints - skipping for now
@@ -110,28 +144,28 @@ function run_ev()
     # @views conserved_checker(Z_t[:,:],dZdt_t[:,:])
 
 
-    #----Renormalization----#
-    meanSqrRenorm = renormalization()
-    zPE = zeroPointEnergy()
+    # #----Renormalization----#
+    # meanSqrRenorm = renormalization()
+    # zPE = zeroPointEnergy()
 
 
-    #----Initial Energy----#
-    totalE = energy(ZED_gl,meanSqrRenorm,zPE)
-    ZedIO = open("data/energies/ZED.dat","w")
-    writedlm(ZedIO,ZED_gl)
-    println("Total initial energy: ",totalE)
+    # #----Initial Energy----#
+    # totalE = energy(ZED_gl,meanSqrRenorm,zPE)
+    # ZedIO = open("data/energies/ZED.dat","w")
+    # writedlm(ZedIO,ZED_gl)
+    # println("Total initial energy: ",totalE)
 
 
-    #----Time Evolution----#
-    @time time_evolve!(ϕ_gl,ψ_gl,ZED_gl,meanSqrRenorm,zPE)
+    # #----Time Evolution----#
+    # @time time_evolve!(ϕ_gl,ψ_gl,ZED_gl,meanSqrRenorm,zPE)
 
 
-    #----Check for Vortices----#
-    vortex_finder(ϕ_gl)
+    # #----Check for Vortices----#
+    # vortex_finder(ϕ_gl)
  
-    #Close data files
-    close(ioϕ)
-    close(ioψ)
+    # #Close data files
+    # close(ioϕ)
+    # close(ioψ)
 
 end